+++ src/titiler/core/titiler/core/factory.py
@@ -357,7 +357,7 @@ class TilerFactory(BaseTilerFactory):
         ):
             """Return the bounds of the COG."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     return {"bounds": src_dst.geographic_bounds}
 
     ############################################################################
@@ -380,7 +380,7 @@ class TilerFactory(BaseTilerFactory):
         ):
             """Return dataset's basic info."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     return src_dst.info()
 
         @self.router.get(
@@ -402,7 +402,7 @@ class TilerFactory(BaseTilerFactory):
         ):
             """Return dataset's basic info as a GeoJSON feature."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     return Feature(
                         type="Feature",
                         geometry=Polygon.from_bounds(*src_dst.geographic_bounds),
@@ -440,19 +440,19 @@ class TilerFactory(BaseTilerFactory):
         ):
             """Get Dataset statistics."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     image = src_dst.preview(
-                        **layer_params,
-                        **image_params,
-                        **dataset_params,
+                        **layer_params.as_dict(),
+                        **image_params.as_dict(),
+                        **dataset_params.as_dict(),
                     )
 
                     if post_process:
                         image = post_process(image)
 
                     return image.statistics(
-                        **stats_params,
-                        hist_options={**histogram_params},
+                        **stats_params.as_dict(),
+                        hist_options=histogram_params.as_dict(),
                     )
 
         # POST endpoint
@@ -491,7 +491,7 @@ class TilerFactory(BaseTilerFactory):
                 fc = FeatureCollection(type="FeatureCollection", features=[geojson])
 
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     for feature in fc:
                         shape = feature.model_dump(exclude_none=True)
                         image = src_dst.feature(
@@ -499,9 +499,9 @@ class TilerFactory(BaseTilerFactory):
                             shape_crs=coord_crs or WGS84_CRS,
                             dst_crs=dst_crs,
                             align_bounds_with_dataset=True,
-                            **layer_params,
-                            **image_params,
-                            **dataset_params,
+                            **layer_params.as_dict(),
+                            **image_params.as_dict(),
+                            **dataset_params.as_dict(),
                         )
 
                         # Get the coverage % array
@@ -514,8 +514,8 @@ class TilerFactory(BaseTilerFactory):
                             image = post_process(image)
 
                         stats = image.statistics(
-                            **stats_params,
-                            hist_options={**histogram_params},
+                            **stats_params.as_dict(),
+                            hist_options=histogram_params.as_dict(),
                             coverage=coverage_array,
                         )
 
@@ -598,15 +598,17 @@ class TilerFactory(BaseTilerFactory):
             """Create map tile from a dataset."""
             tms = self.supported_tms.get(tileMatrixSetId)
             with rasterio.Env(**env):
-                with self.reader(src_path, tms=tms, **reader_params) as src_dst:
+                with self.reader(
+                    src_path, tms=tms, **reader_params.as_dict()
+                ) as src_dst:
                     image = src_dst.tile(
                         x,
                         y,
                         z,
                         tilesize=scale * 256,
-                        **tile_params,
-                        **layer_params,
-                        **dataset_params,
+                        **tile_params.as_dict(),
+                        **layer_params.as_dict(),
+                        **dataset_params.as_dict(),
                     )
                     dst_colormap = getattr(src_dst, "colormap", None)
 
@@ -623,7 +625,7 @@ class TilerFactory(BaseTilerFactory):
                 image,
                 output_format=format,
                 colormap=colormap or dst_colormap,
-                **render_params,
+                **render_params.as_dict(),
             )
 
             return Response(content, media_type=media_type)
@@ -711,7 +713,9 @@ class TilerFactory(BaseTilerFactory):
 
             tms = self.supported_tms.get(tileMatrixSetId)
             with rasterio.Env(**env):
-                with self.reader(src_path, tms=tms, **reader_params) as src_dst:
+                with self.reader(
+                    src_path, tms=tms, **reader_params.as_dict()
+                ) as src_dst:
                     return {
                         "bounds": src_dst.geographic_bounds,
                         "minzoom": minzoom if minzoom is not None else src_dst.minzoom,
@@ -863,7 +867,9 @@ class TilerFactory(BaseTilerFactory):
 
             tms = self.supported_tms.get(tileMatrixSetId)
             with rasterio.Env(**env):
-                with self.reader(src_path, tms=tms, **reader_params) as src_dst:
+                with self.reader(
+                    src_path, tms=tms, **reader_params.as_dict()
+                ) as src_dst:
                     bounds = src_dst.geographic_bounds
                     minzoom = minzoom if minzoom is not None else src_dst.minzoom
                     maxzoom = maxzoom if maxzoom is not None else src_dst.maxzoom
@@ -929,13 +935,13 @@ class TilerFactory(BaseTilerFactory):
             """Get Point value for a dataset."""
 
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     pts = src_dst.point(
                         lon,
                         lat,
                         coord_crs=coord_crs or WGS84_CRS,
-                        **layer_params,
-                        **dataset_params,
+                        **layer_params.as_dict(),
+                        **dataset_params.as_dict(),
                     )
 
             return {
@@ -972,11 +978,11 @@ class TilerFactory(BaseTilerFactory):
         ):
             """Create preview of a dataset."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     image = src_dst.preview(
-                        **layer_params,
-                        **image_params,
-                        **dataset_params,
+                        **layer_params.as_dict(),
+                        **image_params.as_dict(),
+                        **dataset_params.as_dict(),
                         dst_crs=dst_crs,
                     )
                     dst_colormap = getattr(src_dst, "colormap", None)
@@ -994,7 +1000,7 @@ class TilerFactory(BaseTilerFactory):
                 image,
                 output_format=format,
                 colormap=colormap or dst_colormap,
-                **render_params,
+                **render_params.as_dict(),
             )
 
             return Response(content, media_type=media_type)
@@ -1044,9 +1050,9 @@ class TilerFactory(BaseTilerFactory):
                         [minx, miny, maxx, maxy],
                         dst_crs=dst_crs,
                         bounds_crs=coord_crs or WGS84_CRS,
-                        **layer_params,
-                        **image_params,
-                        **dataset_params,
+                        **layer_params.as_dict(),
+                        **image_params.as_dict(),
+                        **dataset_params.as_dict(),
                     )
                     dst_colormap = getattr(src_dst, "colormap", None)
 
@@ -1063,7 +1069,7 @@ class TilerFactory(BaseTilerFactory):
                 image,
                 output_format=format,
                 colormap=colormap or dst_colormap,
-                **render_params,
+                **render_params.as_dict(),
             )
 
             return Response(content, media_type=media_type)
@@ -1103,14 +1109,14 @@ class TilerFactory(BaseTilerFactory):
         ):
             """Create image from a geojson feature."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     image = src_dst.feature(
                         geojson.model_dump(exclude_none=True),
                         shape_crs=coord_crs or WGS84_CRS,
                         dst_crs=dst_crs,
-                        **layer_params,
-                        **image_params,
-                        **dataset_params,
+                        **layer_params.as_dict(),
+                        **image_params.as_dict(),
+                        **dataset_params.as_dict(),
                     )
                     dst_colormap = getattr(src_dst, "colormap", None)
 
@@ -1127,7 +1133,7 @@ class TilerFactory(BaseTilerFactory):
                 image,
                 output_format=format,
                 colormap=colormap or dst_colormap,
-                **render_params,
+                **render_params.as_dict(),
             )
 
             return Response(content, media_type=media_type)
@@ -1179,8 +1185,8 @@ class MultiBaseTilerFactory(TilerFactory):
         ):
             """Return dataset's basic info or the list of available assets."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
-                    return src_dst.info(**asset_params)
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
+                    return src_dst.info(**asset_params.as_dict())
 
         @self.router.get(
             "/info.geojson",
@@ -1202,14 +1208,14 @@ class MultiBaseTilerFactory(TilerFactory):
         ):
             """Return dataset's basic info as a GeoJSON feature."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     return Feature(
                         type="Feature",
                         geometry=Polygon.from_bounds(*src_dst.geographic_bounds),
                         properties={
                             asset: asset_info
                             for asset, asset_info in src_dst.info(
-                                **asset_params
+                                **asset_params.as_dict()
                             ).items()
                         },
                     )
@@ -1226,7 +1232,7 @@ class MultiBaseTilerFactory(TilerFactory):
         ):
             """Return a list of supported assets."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     return src_dst.assets
 
     # Overwrite the `/statistics` endpoint because the MultiBaseReader output model is different (Dict[str, Dict[str, BandStatistics]])
@@ -1258,13 +1264,13 @@ class MultiBaseTilerFactory(TilerFactory):
         ):
             """Per Asset statistics"""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     return src_dst.statistics(
-                        **asset_params,
-                        **image_params,
-                        **dataset_params,
-                        **stats_params,
-                        hist_options={**histogram_params},
+                        **asset_params.as_dict(),
+                        **image_params.as_dict(),
+                        **dataset_params.as_dict(),
+                        **stats_params.as_dict(),
+                        hist_options=histogram_params.as_dict(),
                     )
 
         # MultiBaseReader merged statistics
@@ -1294,23 +1300,23 @@ class MultiBaseTilerFactory(TilerFactory):
         ):
             """Merged assets statistics."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     # Default to all available assets
                     if not layer_params.assets and not layer_params.expression:
                         layer_params.assets = src_dst.assets
 
                     image = src_dst.preview(
-                        **layer_params,
-                        **image_params,
-                        **dataset_params,
+                        **layer_params.as_dict(),
+                        **image_params.as_dict(),
+                        **dataset_params.as_dict(),
                     )
 
                     if post_process:
                         image = post_process(image)
 
                     return image.statistics(
-                        **stats_params,
-                        hist_options={**histogram_params},
+                        **stats_params.as_dict(),
+                        hist_options=histogram_params.as_dict(),
                     )
 
         # POST endpoint
@@ -1349,7 +1355,7 @@ class MultiBaseTilerFactory(TilerFactory):
                 fc = FeatureCollection(type="FeatureCollection", features=[geojson])
 
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     # Default to all available assets
                     if not layer_params.assets and not layer_params.expression:
                         layer_params.assets = src_dst.assets
@@ -1360,16 +1366,17 @@ class MultiBaseTilerFactory(TilerFactory):
                             shape_crs=coord_crs or WGS84_CRS,
                             dst_crs=dst_crs,
                             align_bounds_with_dataset=True,
-                            **layer_params,
-                            **image_params,
-                            **dataset_params,
+                            **layer_params.as_dict(),
+                            **image_params.as_dict(),
+                            **dataset_params.as_dict(),
                         )
 
                         if post_process:
                             image = post_process(image)
 
                         stats = image.statistics(
-                            **stats_params, hist_options={**histogram_params}
+                            **stats_params.as_dict(),
+                            hist_options=histogram_params.as_dict(),
                         )
 
                     feature.properties = feature.properties or {}
@@ -1423,8 +1430,8 @@ class MultiBandTilerFactory(TilerFactory):
         ):
             """Return dataset's basic info."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
-                    return src_dst.info(**bands_params)
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
+                    return src_dst.info(**bands_params.as_dict())
 
         @self.router.get(
             "/info.geojson",
@@ -1446,11 +1453,11 @@ class MultiBandTilerFactory(TilerFactory):
         ):
             """Return dataset's basic info as a GeoJSON feature."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     return Feature(
                         type="Feature",
                         geometry=Polygon.from_bounds(*src_dst.geographic_bounds),
-                        properties=src_dst.info(**bands_params),
+                        properties=src_dst.info(**bands_params.as_dict()),
                     )
 
         @self.router.get(
@@ -1465,7 +1472,7 @@ class MultiBandTilerFactory(TilerFactory):
         ):
             """Return a list of supported bands."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     return src_dst.bands
 
     # Overwrite the `/statistics` endpoint because we need bands to default to the list of bands.
@@ -1497,22 +1504,23 @@ class MultiBandTilerFactory(TilerFactory):
         ):
             """Get Dataset statistics."""
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     # Default to all available bands
                     if not bands_params.bands and not bands_params.expression:
                         bands_params.bands = src_dst.bands
 
                     image = src_dst.preview(
-                        **bands_params,
-                        **image_params,
-                        **dataset_params,
+                        **bands_params.as_dict(),
+                        **image_params.as_dict(),
+                        **dataset_params.as_dict(),
                     )
 
                     if post_process:
                         image = post_process(image)
 
                     return image.statistics(
-                        **stats_params, hist_options={**histogram_params}
+                        **stats_params.as_dict(),
+                        hist_options=histogram_params.as_dict(),
                     )
 
         # POST endpoint
@@ -1551,7 +1559,7 @@ class MultiBandTilerFactory(TilerFactory):
                 fc = FeatureCollection(type="FeatureCollection", features=[geojson])
 
             with rasterio.Env(**env):
-                with self.reader(src_path, **reader_params) as src_dst:
+                with self.reader(src_path, **reader_params.as_dict()) as src_dst:
                     # Default to all available bands
                     if not bands_params.bands and not bands_params.expression:
                         bands_params.bands = src_dst.bands
@@ -1562,16 +1570,17 @@ class MultiBandTilerFactory(TilerFactory):
                             shape_crs=coord_crs or WGS84_CRS,
                             dst_crs=dst_crs,
                             align_bounds_with_dataset=True,
-                            **bands_params,
-                            **image_params,
-                            **dataset_params,
+                            **bands_params.as_dict(),
+                            **image_params.as_dict(),
+                            **dataset_params.as_dict(),
                         )
 
                         if post_process:
                             image = post_process(image)
 
                         stats = image.statistics(
-                            **stats_params, hist_options={**histogram_params}
+                            **stats_params.as_dict(),
+                            hist_options=histogram_params.as_dict(),
                         )
 
                         feature.properties = feature.properties or {}